Remove AtomicTerminalSignalConsumer
